+++ packages/sdk-multisig/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Update dependencies
+
 ## 0.3.0 2026-02-23
 
 ### Changed